Ensure only one thread deletes the database #399

+++ src/HealthChecks.UI/ServiceCollectionExtensions.cs
@@ -14,6 +14,8 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Net.Http;
+using System.Runtime.InteropServices;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Microsoft.Extensions.DependencyInjection
@@ -114,6 +116,8 @@ namespace Microsoft.Extensions.DependencyInjection
             .Services;
         }
 
+        private static bool databaseDeleted = false;
+        private static Mutex deletionMutex = new Mutex();
         static async Task CreateDatabase(IServiceProvider serviceProvider)
         {
             var scopeFactory = serviceProvider.GetRequiredService<IServiceScopeFactory>();
@@ -129,8 +133,14 @@ namespace Microsoft.Extensions.DependencyInjection
                 var settings = scope.ServiceProvider
                     .GetService<IOptions<Settings>>();
 
-                await db.Database.EnsureDeletedAsync();
+                deletionMutex.WaitOne();
+                if (!databaseDeleted)
+                {
+                    await db.Database.EnsureDeletedAsync();
+                    databaseDeleted = true;
+                }
                 await db.Database.MigrateAsync();
+                deletionMutex.ReleaseMutex();
 
                 var healthCheckConfigurations = settings.Value?
                     .HealthChecks?
